add auto-parallel cli tuner
